requirements.txt
# OCR
pytesseract
easyocr
opencv-python
Flask

# ML
torch
torchvision
scikit-learn
numpy
Pillow
albumentations
ultralytics

# NLP / PII
spacy
regex

# API (planned)
fastapi
uvicorn

# Utilities
python-dotenv
tqdm
matplotlib

# Testing
pytest

deepface
tf-keras

qrcode[pil]